@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,8 +34,10 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  flutter_uploader: ^1.2.1
+  flutter_uploader: ^3.0.0-beta.4
   image_picker: ^0.8.5+3
+  shared_preferences: ^2.0.15
+  flutter_local_notifications: ^9.9.0
 
 dev_dependencies:
   flutter_test:
